pcie_host: allow KERNEL_VERSION override
The buildroot that runs on newer machines will fail to build
the driver module if the KERNEL_VERSION cannot be changed
However it is possible to pass KDIR or KERNEL_VERSION to the driver make
so it uses kernel headers for a version different from the host machine.
Right now the Buildroot has to build host binaries for kernel
4.15.0-29-generic, which is not always the running kernel on a buil
machine.

Change-Id: I89f37d37ea6886e0f12213736a523804da85c9f0
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/23384
Reviewed-by: Veerasenareddy Burru <Veerasenareddy.Burru@cavium.com>
Reviewed-by: Yan Markman <ymarkman@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>