pcie_host: remove KERNEL_PATCH_VERSION
Removed KERNEL_PATCH_VERSION as this was originally added to work with
multiple patch revisions of rhel/centos kernels which is not required
anymore. This is creating issues while compiling on regular kernels or
ubuntu releases. Will add more refinely if required in future.

Also removed cav_net_check_skb_frags() as it is not used and inefficient
to go through all fragments for every non-linear packet.

Change-Id: I0409ebe943a2d08503d49ae8c0b569f524ab2be0
Signed-off-by: Veerasenareddy Burru <vburru@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/25239
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Satananda Burla <sburla@cavium.com>