pcie_host: octeontx/2: fix rx buff alloc failure
When a received packet is larger than a DROQ buffer, we allocate a new
large buffer to include all fragments of received packet and push to
stack. If the new buffer allocation fails, currently it keeps running in
while loop as the processed packet length/count is not moved, causing
CPU lock.

Fixed by dropping the Rx packets if the new buffer allocation fails.

Change-Id: Iffefd248f11d6cc007ef9b780252cbcb3ec6630b
Signed-off-by: Veerasenareddy Burru <vburru@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/firewall_offload/generic/host/octeontx/kernel/pcie_host/+/28829
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Satananda Burla <sburla@marvell.com>