host: drivers: introduce facility_dev to note octeon_id
When multiple PFs are there and only single PF is having facility config
then we need a way to identify which device is that. As the order of PCI
detection varies we cannot guarantee that first one is always going to
be the main device for facility.
The facility_dev is introduced to save octeon_id and being used in
facility for triggering interrupt.

Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Change-Id: I30834a5c2df39377aca9807337d042315eb59062
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/pcie_ep/pcie_ep_octeontx/+/50845
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Satananda Burla <sburla@marvell.com>