host: octnic: always use iq ring configuration from firmware
The current host code does not properly handle configurations where the
host driver and firmware have different number of rings configured.  If
the driver tries to use fewer rings than the firmware configured
a mismatch occurs since the host assumes these are contiguous,
however the layout of rings is determined by the firmware.
For now the host driver will take and use whatever configuration
the firmware specifies.
A more complete fix for this is planned in IPBUSW-12515

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I669a509c1d14d52aaca13700299b19657dc9b28e