host: driver: Add sdp_packet_mode param
Add the sdp_packet_mode parameter to the driver to allow selection
of nic mode for NIC style usage, or loop mode for host to OcteonTX2
communication.  If the parameter is not provided, the driver defaults
to nic mode to match the prior behavior.
Usage: sdp_packet_mode=<nic|loop>

IPBUSW-12533
Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: Id01aa7a63c6e5455f85f6c0870159d9da17428eb