host: driver: enable restarting of NIC driver on re-bind
This commit makes several changes to allow PCI driver unbind/bind
to work with the NIC driver.
When unbind/bind is use to disconnect/reconnect the base driver
to Octeon, we need to reconnec the NIC module when the bind
happens.  To do this, we look for an existing handler for
the NIC app, and if that exists set the module status
to OCTEON_MODULE_HANDLER_INIT_LATER so that the NIC
module is initialized on that device when the host/firmware
handshake is complete.

The poll function setup must be done earlier to support bind.
app_mode is called at a different time in the VF driver.

SRIOV is enabled later to ensure that the VFs are created after
the PF initialization is completed.  Without this, the VFs are
created earlier during the PF initialization, and after bind
the VF MSIX interrupts are masked in the MSIX table in the
VF's BAR.  The reason for these interrupts being masked is
not understood.  This same failure mode can be reproduced by
loading the oct_vf_nic driver before the octnic driver.

IPBUSW-16316
Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: If5243c52aed9b4afa5900997e8b7098fbadd9a7c