host: driver: mgmt_net: properly use octeon device index
Properly use Octeon device index when getting a facility conf.  Currently
the mgmt net driver is hard coded to only use Octeon 0, but
mv_get_facility_conf() would succeed when called for index 0 if any
Octeon in the system was ready.
Restiction to Octeon 0 remains, but now driver will properly wait for
Octeon 0 to be ready.

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I926b43c22dffbd0bae4bd95ab3211ffb25a4393a