pcie_ep: Clear OEI interrupt after enabling
Since the OEI interrupt is used for keepalive/heartbeat messages from the
OTX2, interrupts can be received and not cleared before the driver is loaded.
Since the IRQ remains high, further interrupts are not signaled.  We clear
any OEI interrupts that are pending, this way future OEI interrupts are
delivered as expected.

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I579b63a3d79e27e61af86a8d8986c3b6282eded9