pcie_ep: host: Check for valid PCIe port during load
Verify that the PCIe port read early in setup is not all 1s,
which indicates CSR accesses over BAR space not working.
Stop loading the driver if this is detected, and print an error
message.
Continuing to load the driver in this case results in crashes due
to use of invalid register read results.
Remove cleanup on initialization failure, as this cleanup needs to
be done on module removal.  Due to the way the driver waits for FW
READY, the module probe always succeeds, so it is not possible to
just error out and have the load fail to complete.

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I7a18caf46566941065ebc22e17108a3f40d8b4a3