pcie_ep: host: Add octeon_bar_access_valid()
Add the octeon_bar_access_valid() function that returns if BAR based CSR
accesses are returning valid data or not.  Use this to determine
if the OcteonTX is accessible during module removal, and if it is not
then skip shutdown steps that rely on BAR CSR reads to work.
This provides more reliable module removal when the Octeon device is
unexpectedly reset.

IPBUSW-18293
Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I38f897e16b530aba679ea8f6a1d9f971eaed9c08