pcie_ep: Update FW_READY early in boot
Update the FW status register from READY (0x1) to RUNNING (0x2).  This
leaves the status register in a state where an unexpected reset will leave
the register in a state that will prevent the firmware from loading when
it shouldn't. This also tracks the firmware state (running) in the status
register.

Add setting FW status for CN98xx - this was missing in previous change.

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I62e2b0c141850090392732843e178e9d6ea888f8