host: disable IQ interrupt tuning
Disable the periodic IQ interrupt level tuning, as this interferes
with ISM queue depth checking, and should not be needed now that
NAPI support exists.
Update initial IQ interrupt config to match the default config
that would be set with the tuning - 10us time (on CN9KPLUS),
and 0 packets.

The irq tuning code and workqueue should be removed once it is
determined that there is no use case for this code.

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I205f064041b42211e6eef17e09388ae8a87a5b04