host: driver: add conditional compilation for Octeon backports
The Octeon SDK kernel has ethtool related backports that change the
kernel interface.  Add conditional compilation based on build architecture
to allow both x86 and arm64 compiles to work.
This will not work for arm64 kernels that do not have this feature
backported, however there is not a clean fix for this.

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I658aad3018b709edbaf6551bdb9a621901a3803a