host: optim prepare-pci-cmd
The octnet_prepare_pci_cmd() has deep differentiation for octeon-cn83
vs later octeons.
The final placed command is one only, but the procedure does everything
twice - for cn83 and for cn9plus. The API is called per-packet and the
"double prepare" impacts performance for ~1%.

This patch reconsiders the API to prepare the command once only for a
handled device OCTEON_CN83XX_PF_OR_VF or OCTEON_CN9PLUS.

Signed-off-by: Yan Markman <ymarkman@marvell.com>
Change-Id: I19e588ca7b6e79aaa7869e094a4f80f527fd5a2d