host: octnic: Make keepalive/heartbeat more configurable
Increase the frequency of heartbeat intervals from 1 every 5 seconds
to 2 per/second (500ms interval), and change the frequency of checks
for heartbeats to 1/second.  This allows much finer grain control
for how long it will take for a crash to be detected.
(This change is in target pci_ep module.)

Enforce a 15 second minimum timeout at startup, as this ensures
that there is not a timeout between the setting of FW_READY and
the loading of the pcie_ep module that starts the heartbeat.
Change the default miss_limit to 20, which is a reduction from
the about 60 seconds previous default.

Allow setting of the keepalive_miss_limit parameter via sysfs
during runtime.  This allows the miss_limit to be lowered after
the pcie_ep module has been lowered.

NOTE: If you run a new host module with an old pcie_ep module
you will get lots of missed heartbeats due to the different
frequencies.

IPBUSW-21983
Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I7839f54451ea2c22d4b9e030181a804b833b1f86