target: pcie_ep: Create BAR4 mapping index 14 for PTP block
Add a BAR4 mapping of the PTP block using index 14 to enable host access to the
PTP registers. This enables the host PHC driver to directly use the Octeon
PTP registers as a timesource and for the host clock to be synchronized with
Octeon clock.
The use of BAR4 for these register accesses is required due to a HW errata
with the windowed CSR access mechanism.

IPBUSW-19810
Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I412638768d1807de5abb96614ce4e0db280e34a7