host: fix skb_frag_off duplication on different kernel-ver
The skb_frag_off() was officially introduced in skbuf.h
starting from KERNEL-5.4.0.
Most of KERNEL-4.15.0-NN have no this function, but
it has been added into some latest 4.15.0-NN.
For example - NN=55 has no the function but NN=189 does have it.

The "#if LINUX_VERSION_CODE ?? KERNEL_VERSION(4,15,0)" cannot be used
in that case and leads to the build-failure "error: redefinition".
This patch reconsiders the checker next way:
- uses func-name skb_frag_off_local() instead of the skb_frag_off()
- for LNX<K5.4 releases implement skb_frag_off_local()
    e.g. local implementation is used disregarding to global presence
- for LNX>=K5.4 releases define skb_frag_off_local as skb_frag_off

Signed-off-by: Yan Markman <ymarkman@marvell.com>
Change-Id: I7267322972d61ca85b57789f7ccaf440c0f8d7cd