octep_cp_lib: Set correct pem and pf index in mbox message.
mbox messages received from host do not have pem and pf indexes
expected by the firmware. Instead firmware polls on mbox setup on
appropriate pem and pf.
So set the pem and pf index in incoming message headers before
passing on the message to upper layer. Also reset the the pem and
pf index in outgoing messages to the host, since host expects pf
index to be 0 and has no notion of pem index.

This change is required to support multiple pf's on one pem.

Change-Id: I7c2e1e77ee090cf40f72fac06c4c0606650fbaac
Signed-off-by: Abhijit Ayarekar <aayarekar@marvell.com>