l2fwd_pcie_ep: Enable promiscuous mode on data interfaces.
Enable promiscuous mode when interfaces are started.
This is required for forwarding to work when host interface is
added to a bridge. In this case incoming packets from network will
have destination mac address of the bridge.
This change should be removed when control plane supports ndo_ops
to set promiscuous flag on interface. Then host can independently
set the interface in promiscuous mode according to its need.

Change-Id: I0f56d0fe23e87ef6c749b13154f94a012c18c968
Signed-off-by: Abhijit Ayarekar <aayarekar@marvell.com>