l2fwd_pcie_ep: fix offload feature assumptions
offloads can be supported or unsupported by setting
offloads =0/1 in the config file
if offloads are supported, there is always tx metadata
and rx metadata for host traffic irrespective of whether
host enables or disables the offloads.
so to_host rx/tx prep functions in datapath are not changed
when host disables/enables offloads.
when host disables tx offloads, the to_wire tx function
needs to be changed so that mbuf->ol_flags are not enabled
for checksum or tso

Signed-off-by: Satananda Burla <sburla@marvell.com>
Change-Id: I7ec1120658273a7eaa09fe1d3c9134fa6ab6f86b