octboot_net: indicate octeon when host driver is unloaded
When the octboot_net driver is unloaded on host, indicate octeon by
writing to a status bit in host reset status register.
Defined new register at offset 0x88 and bit-0 of this register is used
to indicate host drivers are unloaded.
whenever octboot driver is invoked on octeon, it shall first check this
bit and reset its local state on octeon and start handshake with host
again. Otherwise, host and octeon ends up in different states resulting
in loss of communication.

Signed-off-by: Veerasenareddy Burru <vburru@marvell.com>
Change-Id: Ia51a1a47ac943becf725293c8c6f890a4bed1813