octboot_net: save and restore pci device config
Octeon performs a cold reset, when the octeon firmware is updated and
reset. This is erasing base addresses of all three BAR regions (BAR0,2,4).
When the device reset is complete, octboot_net driver can access config
space but BAR4 is not accessible due to BAR base add config is cleared.

Added octboot_is_pci_bar_addr_reset() to check if BAR base addresses are
cleared; if cleared, restore the PCI config by calling pci_restore_state().
pci state is saved when the driver could read signature successfully.

Also added octboot_is_pci_bar_accessible() to mark device is
unavailable if BAR4 is not accessible, which usually happens when the
Octeon is reset. If the BAR4 is not accessible, do not proceed with device
initialization.

Signed-off-by: Veerasenareddy Burru <vburru@marvell.com>
Change-Id: Id8bcf0cecf4e9cbe8bf33f108eb5bef98e591419