dmadev/dpi_test: spread remote buffers across large region
When a same/small remote memory region is used for a burst of
DPI commands, inbound performance is dropping due to DDR throttling
on remote host (x86 machine). This issue is severely observed
on Xeon servers.

Patch spreads remote buffers across a large DDR region to overcome
the throttling effect.

Signed-off-by: Vamsi Attunuru <vattunuru@marvell.com>
Change-Id: I27164d4b1fe88f371fcf76789e4e61b0c4bb9f27