octep_cp_lib: input SDP RVU PF for SDP register access
Recently the SDP register access was split from the DPI PF BAR0
to separate RVU PFs. Take the second RVU PF as input for accessing
SDP registers such as OEI intr and misc registers.

The DPI PF is now optional to pass into the cp agent. If the DPI PF is
passed in using the --dpi_dev option, the DPI engines will be
initialized by the cp agent, if not, they won't be.

Signed-off-by: Shinas Rasheed <srasheed@marvell.com>
Change-Id: Id0e378257ebea74e4bdfe7656cf3b5a8747726ef