octep_cp_lib: fix host version initial invalid error
Host version is read each time ctrl_mbox_send or receive
is called, but this is inadequate during driver insertion,
since the logic works only because driver tries to send message
and ctrl_mbox_receive is triggered. Also the cp library assumes
that the host version in the BAR4 will be zeroed out as initial value.
This can be untrue, so check the initial value "0xa5a5a5a5a5a5a5a5"
and take decision.

Signed-off-by: Shinas Rasheed <srasheed@marvell.com>
Change-Id: Ifed4a6b25877d02ae19a93a64229c3ac4f81399a